refactor tasks into separate modules
